Dependency flip flop
