+++ crates/linear_algebra/Cargo.toml
@@ -9,13 +9,11 @@ homepage.workspace = true
 approx = { workspace = true, optional = true }
 nalgebra = { workspace = true }
 num-traits = { workspace = true }
-path_serde = { workspace = true, optional = true }
 serde = { workspace = true, optional = true }
 simba = { workspace = true }
 
 [features]
 serde = ["dep:serde"]
-path_serde = ["dep:path_serde"]
 approx = ["dep:approx"]
 
-full = ["serde", "path_serde", "approx"]
+full = ["serde", "approx"]
